Kill all children when running `async_flush_jobs`
When flushing all jobs, all child processes must be sent a TERM signal,
not just the immediate child processes. Previously the `_async_job` was
killed resulting in no output from the job, however, anything executed
inside the job would not receive a signal to stop and be allowed to
complete whatever it was doing.