Fix quoting issue in _async_job
Typeset (-p) prints with (minimal) shell quoting equivalent to (q+),
this is insufficient when the output contains a string such as '\n'
which should not be interpreted as a newline. By manually printing with
a higher level of quoting, we ensure that the output is interpreted as
it should be.

This change passes `test_async_job_print_matches_input_exactly`.