Fix async_start_worker when noclobber is set
This commit fixes a clobbering issue of errfd because it was initialized
as an integer (and thus zero value). Instead of always allowing
clobbering, we initialize it as -1 which makes zsh happy and protects us
from accidentally clobbering something in the future.

The error fixed was:

	can't clobber parameter errfd containing file descriptor 0

Possible fix for clobbering issue mentioned in #35.