Force zpty fd to behave by wrapping message in newlines
There seems to be a chance that, when the zpty simultaneously outputs
and receives data, the output data is lost.

Introducing newlines seems to fix the issue. Simply introducing trailing
newlines did not. Reason is not entirely clear, perhaps there is
something special about the zpty fd or perhaps this is a property of
`zpty -r` / `read`.