Add sleep duration that configurable via ENV
This sleep duration will be used to give a leeway time for some
terminals that haven't loaded all the necessary scripts when the async
job is being dispatched. In some cases, it will make the async job not
working properly or even not running at all.

We set the default sleep duration as `0` to keep backward compatibility
with the previous version and it can be changed through ENV.